Initial Check-in
Based on the Udemy class - Complete Python Web Course: Build 8 Python
Web Apps (https://bit.ly/2tI0z5e)

This is an app that tracks prices on websites periodically to see if
they come under a price set by the user. This app runs on Flask, using
MongoDB as the data store. I added unit testing with pytest and unittest
(esp. mock). As a rule, the project can only be checked in if the code
passes unit testing and has good coverage (100%?). MongoDB calls
were a challenge to patch with mock; I didn't want to use a special mock
lib just for Mongo though. It would be yet another dependency, another
maintainer to trust. I did keep secrets in a .env file and loaded that with
python-dotenv; it is a challenge to find an authoritative source for best
practices around app config and scaffolding.

Next up is adding the web interface (and unit tests for that).

I probably need to add comments but that is not my priority right now. I
just want to practice more Python. My major goals are to put this app in
a Docker container and put the price update work into Celery.